Move assets to src to stop vite warnings, add scale requirement for fear and dread
