+++ src/assets/death.png
